Intellecto-Edtech-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.15.1  # Updated to latest version
motor==3.7.1     # Updated to latest version compatible with pymongo
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic_settings==2.10.1
beanie==2.0.0  # Added ODM for MongoDB